Add files via upload
This program code will help in calculating area for different 2D shapes. Area calculation can be done for a circle, square, triangle, rectangle, pentagon, or hexagon. In case of a pentagon or hexagon, the program considers it as symmetric. The program will ask for necessary user input on the shape and show the area.