Add files via upload
When there are so many answer sheets to check for a descriptive type question, a teacher needs to spend a lot to time to evaluate all the answers for each student. This program attempts to build a logic, by which it awards marks to a student's descriptive type answer. It checks for certain keywords that need to be present in the answer. Based on suggested criteria, it awards marks to each question's answer. This program has the potential to save significant time for a teacher, from a repetitive evaluation process, without compromising on evaluation.